♻️ Refactor: Donation 컴포넌트 style 파일 코드 오류 개선 및 불필요한 파일 삭제
불필요한 import 제거 import 경로 @로 수정 style.js 파일 styles.js로 변경 css 파일 내부 코드 오류 수정 Resolves: #178
