+++ src/App.jsx
@@ -11,7 +11,6 @@ import Landing from "./pages/Landing";
 import List from "./pages/List";
 import Mypage from "./pages/Mypage";
 import NotFound from "./pages/NotFound/index.jsx";
-import Testpage from "./pages/Testpage/index.jsx";
 
 function App() {
 	const lenisRef = useRef(null);
@@ -46,7 +45,6 @@ function App() {
 				<Route path="/" element={<DefaultLayout />}>
 					<Route path="/list" element={<List />} />
 					<Route path="/mypage" element={<Mypage />} />
-					<Route path="/testpage" element={<Testpage />} />
 					<Route path="/donation-detail/:id" element={<DonationDetail />} />
 				</Route>
 				<Route path="*" element={<NotFound />} />
