--- src/mocks/donations.js
@@ -1,78 +1,0 @@
-export const DONATIONS = [
-	{
-		status: true,
-		deadline: "2025-04-21T00:00:00.000Z",
-		createdAt: "2025-04-15T09:53:35.947Z",
-		receivedDonations: 500000,
-		targetDonation: 1000000,
-		idolId: 0,
-		subtitle: "강남역 10번 출구",
-		title: "민지 생일 축하",
-		id: 0,
-		idol: {
-			totalVotes: 1000,
-			profilePicture: "/mocks/mock01.png",
-			group: "뉴진스",
-			gender: "female",
-			name: "민지",
-			id: 0,
-		},
-	},
-	{
-		status: true,
-		deadline: "2025-12-25T00:00:00.000Z",
-		createdAt: "2025-06-20T14:22:15.123Z",
-		receivedDonations: 2000000,
-		targetDonation: 2000000,
-		idolId: 1,
-		subtitle: "코엑스몰 메가박스 앞",
-		title: "원영이 20번째 생일",
-		id: 1,
-		idol: {
-			totalVotes: 800,
-			profilePicture: "/mocks/mock02.png",
-			group: "아이브",
-			gender: "female",
-			name: "장원영",
-			id: 1,
-		},
-	},
-	{
-		status: false,
-		deadline: "2025-08-31T00:00:00.000Z",
-		createdAt: "2025-05-01T10:15:45.789Z",
-		receivedDonations: 300000,
-		targetDonation: 1500000,
-		idolId: 2,
-		subtitle: "홍대입구역 9번 출구",
-		title: "채원 팬미팅 서포트",
-		id: 2,
-		idol: {
-			totalVotes: 750,
-			profilePicture: "/mocks/mock03.png",
-			group: "르세라핌",
-			gender: "female",
-			name: "채원",
-			id: 2,
-		},
-	},
-	{
-		status: true,
-		deadline: "2026-01-15T00:00:00.000Z",
-		createdAt: "2025-07-10T08:45:30.456Z",
-		receivedDonations: 900000,
-		targetDonation: 3000000,
-		idolId: 3,
-		subtitle: "잠실 롯데월드몰 광장",
-		title: "블랙핑크 제니 지하철 광고",
-		id: 3,
-		idol: {
-			totalVotes: 650,
-			profilePicture: "/mocks/mock04.png",
-			group: "블랙핑크",
-			gender: "female",
-			name: "제니",
-			id: 3,
-		},
-	},
-];
