--- src/mocks/idols.js
@@ -1,102 +1,0 @@
-export const IDOLS = [
-	{
-		voteId: 1,
-		idol: {
-			totalVotes: 1000,
-			profilePicture: "/mocks/mock01.png",
-			group: "뉴진스",
-			gender: "female",
-			name: "민지",
-			id: 0,
-		},
-	},
-	{
-		voteId: 2,
-		idol: {
-			totalVotes: 800,
-			profilePicture: "/mocks/mock02.png",
-			group: "아이브",
-			gender: "female",
-			name: "장원영",
-			id: 1,
-		},
-	},
-	{
-		voteId: 3,
-		idol: {
-			totalVotes: 750,
-			profilePicture: "/mocks/mock03.png",
-			group: "르세라핌",
-			gender: "female",
-			name: "채원",
-			id: 2,
-		},
-	},
-	{
-		voteId: 4,
-		idol: {
-			totalVotes: 650,
-			profilePicture: "/mocks/mock04.png",
-			group: "블랙핑크",
-			gender: "female",
-			name: "제니",
-			id: 3,
-		},
-	},
-	// 추가
-	{
-		voteId: 5,
-		idol: {
-			totalVotes: 1000,
-			profilePicture: "/mocks/mock01.png",
-			group: "뉴진스",
-			gender: "female",
-			name: "민지",
-			id: 4,
-		},
-	},
-	{
-		voteId: 6,
-		idol: {
-			totalVotes: 1000,
-			profilePicture: "/mocks/mock01.png",
-			group: "뉴진스",
-			gender: "female",
-			name: "민지",
-			id: 5,
-		},
-	},
-	{
-		voteId: 7,
-		idol: {
-			totalVotes: 1000,
-			profilePicture: "/mocks/mock01.png",
-			group: "뉴진스",
-			gender: "female",
-			name: "민지",
-			id: 6,
-		},
-	},
-	{
-		voteId: 8,
-		idol: {
-			totalVotes: 1000,
-			profilePicture: "/mocks/mock01.png",
-			group: "뉴진스",
-			gender: "female",
-			name: "민지",
-			id: 7,
-		},
-	},
-	{
-		voteId: 9,
-		idol: {
-			totalVotes: 1000,
-			profilePicture: "/mocks/mock01.png",
-			group: "뉴진스",
-			gender: "female",
-			name: "민지",
-			id: 8,
-		},
-	},
-];
